+++ custom_components/ble_led_pixel/color.py
@@ -95,6 +95,7 @@ class BleLedPixelColorBase(TextEntity, RestoreEntity):
         self._attr_device_info = DeviceInfo(
             identifiers={(DOMAIN, address)},
             name=name,
+            manufacturer="LED_BLE",
             model="LED Pixel Panel",
             sw_version="1.0",
         )