+++ custom_components/ble_led_pixel/light.py
@@ -77,6 +77,7 @@ class BleLedPixelColorLight(LightEntity, RestoreEntity):
         self._attr_device_info = DeviceInfo(
             identifiers={(DOMAIN, address)},
             name=name,
+            manufacturer="LED_BLE",
             model="LED Pixel Panel",
             sw_version="1.0",
         )