+++ custom_components/ble_led_pixel/sensor.py
@@ -101,6 +101,7 @@ class BleLedPixelSensor(SensorEntity):
         self._attr_device_info = DeviceInfo(
             identifiers={(DOMAIN, address)},
             name=name,
+            manufacturer="LED_BLE",
             model="LED Pixel Panel",
             sw_version="1.0",
         )